ADM-601:[frontend]feat: modify the component structure of config-page
