+++ frontend/src/components/Metrics/ConfigStep/BasicInfo/RequiredMetrics/style.tsx
@@ -4,4 +4,5 @@ import { FormControl } from '@mui/material'
 export const RequireDataSelections = styled(FormControl)({
   width: '100%',
   paddingBottom: '1rem',
+  marginTop: '1.25rem',
 })
