+++ frontend/src/components/Metrics/ConfigStep/style.tsx
@@ -1,29 +1,5 @@
-import { styled, css } from '@mui/material/styles'
-import { FormControlLabel, TextField } from '@mui/material'
-import { theme } from '@src/theme'
+import { styled } from '@mui/material/styles'
 
 export const ConfigStepWrapper = styled('div')({
   width: '100%',
-})
-
-export const ProjectNameInput = styled(TextField)({
-  width: '100%',
-})
-
-export const StyledFormControlLabel = styled(FormControlLabel)`
-  ${css`
-    ${theme.breakpoints.down('sm')} {
-      & .MuiFormControlLabel-label {
-        font-size: 0.8rem;
-      }
-    }
-  `}
-`
-
-export const CollectionDateLabel = styled('div')({
-  width: '100%',
-  color: 'rgba(0, 0, 0, 0.6)',
-  fontSize: '0.8rem',
-  lineHeight: '2em',
-  boxSizing: 'border-box',
 })
